requirements.txt
yfinance
pandas
numpy
matplotlib
python-telegram-bot==20.3
flask
